[feat][coordinator] Add DropSchema function, reduce coordinator memory usage.
Coordinator orignally store DingoCommonId, which is 24 bytes large.
If there are 10,000 table in one schema, this schema will use 2.3MB
memory for a single value in std::map.
After this commit, coordinator will not store any DingoCommonId but only
entity_id, which is only 8 bytes, this will reduce the memory to one
third of before.

Signed-off-by: Ketor <d.ketor@gmail.com>